Great implementation of local music import.
May just adjust the css file to make dialog look better later.

+++ src/app/excuseme/util/ImportMusicTask.java
@@ -5,7 +5,8 @@ import javafx.concurrent.Task;
 public abstract class ImportMusicTask<V> extends Task<V> {
 	
 	public void updateProgress(int progress, int maxProgress){
-		updateProgress(progress, maxProgress);
+		//必须加入(long)否则调用的是当前的updateProgress，会陷入死循环
+		updateProgress((long)progress, (long)maxProgress);
 	}
 	
 }
